Fix the decoding of filter settings cookie
Previously, the use of json_decode without a second paramter meant that an
stdClass was returned, which does not allow access to properties via the index
operator. Instead, we now use json_decode(..., true) to return an associative
array.

In order to prevent any type shenanigans, we also ensure whether the returned
value is an array and if not, replace it with an empty one.